in C# you can't return score > 1.0
in the C# Bot Builder SDK you can not return a score greater than 1.0 or less than 0.0

This code in the C# SDK will throw an exception.
https://github.com/Microsoft/BotBuilder/blob/497252e8d9949be20baa2cebaa6ce56de04461cf/CSharp/Library/Microsoft.Bot.Builder/Scorables/Scorables.cs#L375

I was confused by this statement in the documentation than the score can be greater than 1.0. I don't know if it really possible in Node.js, but in C# it's not.